Settled on the fastest scheme
